+++ src/Randomizer/CaseyNoInstantPatch.js
@@ -1,5 +1,4 @@
-/* eslint import/no-webpack-loader-syntax: off */
-import patchText from '!array-loader!./tables/casey_no_instant_patch.txt';
+import patchText from './tables/casey_no_instant_patch.txt';
 import { utils } from 'randomtools-js';
 
 const CaseyNoInstantPatch = {