+++ src/Randomizer/ItemDropPoolPatch.js
@@ -1,5 +1,4 @@
-/* eslint import/no-webpack-loader-syntax: off */
-import patchText from '!array-loader!./tables/item_drop_pool_patch.txt';
+import patchText from './tables/item_drop_pool_patch.txt';
 import { utils } from 'randomtools-js';
 
 const ItemDropPoolPatch = {