+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## v53.x / v54 (In progress)
+- PC sprites: Added Jean-Luc Picard, Dianna Troi, William Riker, and Worf.
+
 ## v53 (2024-04-14)
 - Ancient Cave: Zoo room NPCs removed to prevent hardlocks.
 - Challenges: New challenge mode "Ancient Cave: Short Skips" added.